requirements.txt
pandas
numpy
scikit-learn
streamlit
langchain
langchain-groq
python-dotenv
